requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
pdfplumber
tqdm
pinecone-client>=2.2.0
openai
sentence-transformers
nltk
faiss-cpu
